Final project version

+++ app/Http/Controllers/CommentController.php
@@ -3,20 +3,21 @@
 namespace App\Http\Controllers;
 
 use Illuminate\Support\Facades\Auth;
-use App\Models\Post;
 use App\Models\Comment;
 use Illuminate\Http\Request;
 
 class CommentController extends Controller
 {
-    public function store(Post $post, Request $request)
+    public function isCommentOwnerOrAdmin($comment)
+    {
+        return ($comment->user_id != Auth::id() && Auth::user()->admin_role != 1) ? false : true;
+    }
+
+    public function store(Request $request)
     {
         $request->validate([
             'comment_content' => 'required'
         ]);
-
-        $post = Post::find($request->post_id)->first();
-        if (!$post->id || Auth::guest()) return redirect('/', 400)->with('success', 'Comment not added. Error occured.');
 
         $comment = Comment::create($request->all());
         return redirect('/post/' . strval($comment->post_id))
@@ -30,8 +31,10 @@ class CommentController extends Controller
         ]);
     }
 
-    public function destroy(Post $post, Comment $comment)
+    public function destroy(Comment $comment)
     {
+        if (!$this->isCommentOwnerOrAdmin($comment)) return redirect('/');
+
         $postID = $comment->post_id;
         $comment->delete();
 
@@ -42,11 +45,13 @@ class CommentController extends Controller
 
     public function update(Request $request, Comment $comment)
     {
+        if (!$this->isCommentOwnerOrAdmin($comment)) return redirect('/');
+
         $request->validate([
             'comment_content' => 'required'
         ]);
-
         $comment->update($request->all());
+
         return redirect('/post/' . strval($comment->post_id))
             ->with('success', 'Comment updated successfully.');
     }
