+++ app/Http/Controllers/PostController.php
@@ -11,6 +11,11 @@ use Illuminate\Http\Request;
 
 class PostController extends Controller
 {
+    public function isPostOwnerOrAdmin($post)
+    {
+        return ($post->user_id != Auth::id() && Auth::user()->admin_role != 1) ? false : true;
+    }
+
     public function index()
     {
         $title = "Posts of all users";
@@ -24,7 +29,8 @@ class PostController extends Controller
 
         return view('posts.main', [
             'posts' => $posts,
-            'title' => $title
+            'title' => $title,
+            'page' => 'main'
         ]);
     }
 
@@ -43,7 +49,8 @@ class PostController extends Controller
 
         return view('posts.main', [
             'posts' => $posts,
-            'title' => $title
+            'title' => $title,
+            'page' => 'feed'
         ]);
     }
 
@@ -65,6 +72,7 @@ class PostController extends Controller
             'posts' => $posts,
             'title' => $username,
             'user' => $usr,
+            'page' => 'user',
             'following' => $following
         ]);
     }
@@ -104,6 +112,8 @@ class PostController extends Controller
 
     public function destroy(Post $post)
     {
+        if (!$this->isPostOwnerOrAdmin($post)) return redirect('/');
+
         $userusername = $post->user->username;
         $post->delete();
 
@@ -114,11 +124,13 @@ class PostController extends Controller
 
     public function update(Request $request, Post $post)
     {
+        if (!$this->isPostOwnerOrAdmin($post)) return redirect('/');
+
         $request->validate([
             'post_content' => 'required'
         ]);
-
         $post->update($request->all());
+
         return redirect('/post/' . strval($post->id))
             ->with('success', 'Post updated successfully.');
     }
